fix: исправляет типизацию
