+++ packages/client/src/services/api/APIClientRepository.ts
@@ -1,19 +1,16 @@
 import { request } from "./apiRequest";
-import { TUserRepository, UserDTO } from "./types";
+import { ThemeType, TUserRepository, UserDTO } from "./types";
 
 export class APIClientRepository implements TUserRepository {
   async getCurrent(): Promise<UserDTO> {
     return request.get<UserDTO>("auth/user");
   }
 
-  // TODO: добавить типы
-  async getForumThemes() {
-    const response = await request.get("/forum/themes/", {
+  async getForumThemes(): Promise<ThemeType> {
+    const response = await request.get<{ data: ThemeType }>("/forum/themes/", {
       baseURL: __API_BASEURL__,
     });
 
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
     return response.data;
   }
 }
