+++ packages/client/src/services/api/types.ts
@@ -13,11 +13,9 @@ export type APIError = {
 export type OAuthServiceRequestData = {
   redirect_uri: string;
 };
-
 export type OAuthServiceResponse = {
   service_id: string;
 };
-
 export type OAuthLoginRequestData = {
   code: string;
   redirect_uri: string;
@@ -27,12 +25,6 @@ export type LoginRequestData = {
   login: string;
   password: string;
 };
-
-export type PasswordChangeRequest = {
-  oldPassword: string;
-  newPassword: string;
-};
-
 export type RegisterRequestData = {
   first_name: string;
   second_name: string;
@@ -40,6 +32,10 @@ export type RegisterRequestData = {
   email: string;
   password: string;
   phone: string;
+};
+export type PasswordChangeRequest = {
+  oldPassword: string;
+  newPassword: string;
 };
 
 export type UpdateUserRequestData = {
@@ -50,9 +46,7 @@ export type UpdateUserRequestData = {
   email: string;
   phone: string;
 };
-
 export type UserDTO = z.infer<typeof apiSchema.UserDTO>;
-
 export type User = {
   id: number;
   login: string;
@@ -72,23 +66,19 @@ export type UserToLeboardData = {
   winsAmount: number;
   [RATING_FIELD_NAME]: number;
 };
-
 export type UserToLeboardExtData = {
   data: UserToLeboardData;
   ratingFieldName: string;
   teamName: string;
 };
-
 export type GetTeamLeboardData = {
   ratingFieldName: string;
   cursor: number;
   limit: number;
 };
-
 export type GetTeamLeboardResponse = {
   data: UserToLeboardData;
 };
-
 export type LeaderboardUserData = {
   position: number;
   imgUrl: string | undefined;
@@ -101,14 +91,50 @@ export type LeaderboardUserData = {
 
 export type TUserRepository = {
   getCurrent(): Promise<UserDTO>;
-  // TODO: добавить типы
-  getForumThemes(): Promise<any>;
+  getForumThemes(): Promise<ThemeType>;
 };
-
 export type TUserService = {
   getCurrentUser(): Promise<User>;
-  // TODO: добавить типы
-  getForumThemes(): Promise<any>;
+  getForumThemes(): Promise<ThemeType>;
+};
+// форум
+export type ThemeType = {
+  id: number;
+  user_id: number;
+  title: string;
+  createdAt: Date;
+};
+export type MessageType = {
+  id: number;
+  theme_id: number;
+  user_id: number;
+  user_display_name: string;
+  user_avatar: string;
+  message: string;
+  parent_message_id: number | null;
+  parent_message_text: string | null;
+  createdAt: Date;
+};
+export type ReactionsType = {
+  id: number;
+  theme_id: number;
+  message_id: number;
+  user_id: number;
+  reaction: string;
+  createdAt: Date;
+};
+
+export type RequestTheme = Omit<ThemeType, "id" | "createdAt">;
+export type RequestMessage = Omit<MessageType, "id" | "createdAt">;
+export type RequestReaction = Omit<ReactionsType, "id" | "createdAt">;
+
+export type ForumState = {
+  isLoading: boolean;
+  error: string | null;
+  isSuccess: boolean;
+  themes: ThemeType[];
+  currentMessages: MessageType[];
+  currentReactions: ReactionsType[];
 };
 
 export type AsyncThunkConfig = {
